Refactor code and add Qodana configuration
Several updates have been made in the codebase, starting with removing unused imports and correcting the navigation path in the signin component. The AppModule path in the main file was also adjusted for clarity. NgModel was fixed in reset component to make it two-way data binding. A Qodana configuration file was introduced for better code analysis. Additionally, a redundant import in the app component was removed.